feat : Complete exposing Subscriptions and Commissions data.

+++ foodeals-admin-api/src/main/java/net/foodeals/contract/application/DTo/upload/ContractSubscriptionDto.java
@@ -16,7 +16,7 @@ import java.util.stream.Collectors;
 @Data
 public class ContractSubscriptionDto {
 
-    private Float duration;
+    private Integer duration;
 
     private Float annualPayment;
